+++ src/router/index.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react';
 import firebase from '../firebase';
 import GameContainer from '../components/GameContainer';
-import Login from '../components/Login';
 import { Route, BrowserRouter, Switch } from 'react-router-dom';
 import * as actions from '../actions';
 import { connect } from 'react-redux';
-import PublicRoute from '../components/PublicRoute';
+// import PublicRoute from '../components/PublicRoute';
 // import PrivateRoute from '../components/PrivateRoute';
 
 class Root extends Component {
@@ -32,7 +31,6 @@ class Root extends Component {
           <div className="w-100 bg-light-gray min-vh-100">
             <Switch>
               <Route path="/" exact component={GameContainer} />
-              <PublicRoute path="/login" component={Login} />
               <Route render={() => <h3>No Match</h3>} />
             </Switch>
           </div>
