+++ src/styles/index.css
@@ -4,28 +4,6 @@ body {
   font-family: sans-serif;
 }
 
-.page-title {
-  text-align: center;
-  padding: 2rem 0;
-  margin: 0;
-}
-
-.callout-auth {
-  background-color: #fafafa;
-  border: 1px solid #eeeeee;
-  padding: 2rem;
-  margin: 0 auto;
-  max-width: 20rem;
-  text-align: center;
-}
-
-.callout-auth .button {
-  background-color: lightblue;
-  color: black;
-  padding: 0.5rem;
-  cursor: pointer;
-}
-
 .back-button {
   top: 1.7rem;
   left: 1.7rem;
